Code Cleanup and minor features
C# version 9 was released, as so some cleanup was done that reflects some of the new features from C#9.

Logging class was changed to internal access only.

ReadOnlyGame now supports getting buildables with a transform or instanceId (instanceId may be incorrect due to the split between barricades/structures, this needs to be checked)

Plugin's logo is now included.

Updated to Harmony 2.0.4

Changed the method hooked on the event `onCommenceShutdown` to force saving everything instead of just saving base clustering's stuff.

Added 3 new methods to get clusters with specific elements in them.