Increase BuildableDirectory speed
Instead of using a single list with all buildables, 2 Dictionaries are used, with the InstanceId of each buildable used as the key. This provides a faster search time, but slows down iteration and adding multiple elements.
Also changed the behavior of the deferred destroy. Instead of removing the elements on a separate thread, they are all removed/added on main thread only. This prevents any CollectionModified exceptions, or undefined behaviour, due to multi-thread access.
XML Documentation generation to an XML file has been enabled.
Removed the BuildableSpawns patch, as nelson's OnBarricadeSpawned/OnStructureSpawned now happen correctly, when both datas and drops are synced.